+++ src/openTelemetry/providers/index.ts
@@ -1,3 +1,2 @@
 export * from './logs';
-export * from './metrics';
 export * from './traces';